Added a second view "summary.html.erb" and implemented a logic to render the total amount of hours attended per student => the summary was done in the model by creating a class method "self.total_time_attended"
Added a admin function to be able to delete a database record from the table "attendances". This was done by using http DELETE, the method is called by pressing a button to the said record, while passing its ID so the controller action can correctly know what exact record to delete.